Moved close-icon.svg to assets/images/common folder
